requirements.txt
matplotlib==3.5.1
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.1.0
torch==1.10.2
torchvision==0.11.3
